--- specs/payment/flutterwave/webhook_charge_completed/schema.json
@@ -1,89 +1,0 @@
-{
-  "spec_version": "1.0",
-  "provider_api_version": "v4",
-  "capability": "webhook_charge_completed",
-  "capability_type": "webhook",
-  "status": "draft",
-  "currency": ["GHS", "KES", "NGN", "RWF", "SLE", "TZS", "UGX", "XAF", "XOF", "ZAR", "ZMW"],
-  "auth": {
-    "type": "api_key",
-    "location": "header",
-    "field": "verif-hash",
-    "format": "{token}",
-    "env_var": "FLUTTERWAVE_WEBHOOK_SECRET"
-  },
-  "endpoint": {
-    "method": "POST",
-    "url": "{your_webhook_url}"
-  },
-  "example_prompt": "Receive a charge.completed webhook from Flutterwave, verify the HMAC-SHA256 signature, then fulfil the customer order.",
-  "input_schema": {
-    "type": "object",
-    "required": ["type", "id", "timestamp", "data"],
-    "properties": {
-      "type": {
-        "type": "string",
-        "const": "charge.completed",
-        "description": "Event type. Always \"charge.completed\" for this webhook."
-      },
-      "id": {
-        "type": "string",
-        "description": "Unique webhook event id (e.g. \"wbk_W5p6ktwU0jQ8RO4By860\"). Use it to deduplicate retried deliveries. NOTE: some Flutterwave v4 webhook payloads use \"webhook_id\" instead of \"id\" — handle both."
-      },
-      "webhook_id": {
-        "type": "string",
-        "description": "Alternative top-level event id field — Flutterwave is inconsistent across event types (refund.completed uses webhook_id, charge.completed historically uses id). Treat as a fallback for id."
-      },
-      "timestamp": {
-        "type": "integer",
-        "description": "Unix epoch timestamp in milliseconds when the event was emitted."
-      },
-      "data": {
-        "type": "object",
-        "description": "Charge resource details.",
-        "properties": {
-          "id": { "type": "string", "description": "Charge id. Re-fetch via GET /charges/{id} to verify server-side." },
-          "status": { "type": "string", "description": "succeeded | pending | failed | voided" },
-          "amount": { "type": "number" },
-          "currency": { "type": "string" },
-          "reference": { "type": "string" },
-          "customer_id": { "type": "string" },
-          "payment_method_id": { "type": "string" },
-          "payment_method": {
-            "type": "object",
-            "description": "Snapshot of the payment method used.",
-            "additionalProperties": true
-          },
-          "processor_response": {
-            "type": "object",
-            "description": "Provider-level processor response (codes, raw messages).",
-            "additionalProperties": true
-          },
-          "order_id": { "type": ["string", "null"] },
-          "meta": { "type": ["object", "null"], "additionalProperties": true },
-          "created_datetime": { "type": "string", "format": "date-time" },
-          "updated_datetime": { "type": "string", "format": "date-time" }
-        },
-        "additionalProperties": true
-      }
-    }
-  },
-  "response_schema": {
-    "type": "object",
-    "description": "Your endpoint must respond with HTTP 200 and an (optionally empty) JSON body within 60 seconds.",
-    "properties": {}
-  },
-  "error_schema": {
-    "type": "object",
-    "description": "Webhooks do not return errors to Flutterwave. Any non-2xx response triggers retries.",
-    "properties": {}
-  },
-  "gotchas": [
-    "Flutterwave signs webhook payloads with HMAC-SHA256 using the secret hash configured in your dashboard. Compute HMAC-SHA256 of the raw request body and compare with the 'flutterwave-signature' header. Reject mismatches. Respond with HTTP 200 within 60 seconds — Flutterwave retries failed webhooks 3 times at 30-minute intervals.",
-    "Webhook payloads contain a top-level 'type' field (e.g. 'charge.completed', 'refund.completed') and a 'data' object with the resource details. Always verify the transaction by calling GET /charges/{id} server-side before fulfilling — never trust webhook data alone.",
-    "Use the raw request body (bytes/string) when computing HMAC — re-serialising parsed JSON will reorder keys and break verification.",
-    "Use a constant-time comparison (e.g. node:crypto timingSafeEqual) when comparing signatures to avoid timing-side-channel attacks.",
-    "Deduplicate on the top-level event id — Flutterwave may redeliver the same event after a transient failure on your side.",
-    "data.status is not always 'succeeded' on a charge.completed event. Check the status field explicitly before fulfilling; failed/voided charges also generate this event."
-  ]
-}
