--- specs/payment/flutterwave/webhook_refund_completed/schema.json
@@ -1,80 +1,0 @@
-{
-  "spec_version": "1.0",
-  "provider_api_version": "v4",
-  "capability": "webhook_refund_completed",
-  "capability_type": "webhook",
-  "status": "draft",
-  "currency": ["GHS", "KES", "NGN", "RWF", "SLE", "TZS", "UGX", "XAF", "XOF", "ZAR", "ZMW"],
-  "auth": {
-    "type": "api_key",
-    "location": "header",
-    "field": "verif-hash",
-    "format": "{token}",
-    "env_var": "FLUTTERWAVE_WEBHOOK_SECRET"
-  },
-  "endpoint": {
-    "method": "POST",
-    "url": "{your_webhook_url}"
-  },
-  "example_prompt": "Receive a refund.completed webhook from Flutterwave, verify the HMAC-SHA256 signature, then mark the associated order as refunded.",
-  "input_schema": {
-    "type": "object",
-    "required": ["type", "webhook_id", "timestamp", "data"],
-    "properties": {
-      "type": {
-        "type": "string",
-        "const": "refund.completed",
-        "description": "Event type. Always \"refund.completed\" for this webhook."
-      },
-      "webhook_id": {
-        "type": "string",
-        "description": "Unique webhook event id (e.g. \"wh_abc123\"). Use it to deduplicate retried deliveries."
-      },
-      "timestamp": {
-        "type": "integer",
-        "description": "Unix epoch timestamp (int64) when the event was emitted (e.g. 1701619661)."
-      },
-      "data": {
-        "type": "object",
-        "description": "Refund resource details.",
-        "required": ["id", "charge_id", "status"],
-        "properties": {
-          "id": { "type": "string", "description": "Refund id (e.g. \"ref_ABC123XYZ\")." },
-          "charge_id": { "type": "string", "description": "Id of the original charge being refunded." },
-          "amount_refunded": { "type": "number", "description": "Amount refunded (may be less than the original charge for partial refunds)." },
-          "reason": {
-            "type": "string",
-            "enum": ["duplicate", "fraudulent", "requested_by_customer", "expired_uncaptured_charge"],
-            "description": "Reason that was supplied when the refund was created."
-          },
-          "status": {
-            "type": "string",
-            "enum": ["pending", "requires_action", "succeeded", "failed", "cancelled", "completed", "new"],
-            "description": "Current refund lifecycle state."
-          },
-          "meta": { "type": ["object", "null"], "additionalProperties": { "type": "string" } },
-          "created_datetime": { "type": "string", "format": "date-time" }
-        }
-      }
-    }
-  },
-  "response_schema": {
-    "type": "object",
-    "description": "Your endpoint must respond with HTTP 200 and an (optionally empty) JSON body within 60 seconds.",
-    "properties": {}
-  },
-  "error_schema": {
-    "type": "object",
-    "description": "Webhooks do not return errors to Flutterwave. Any non-2xx response triggers retries.",
-    "properties": {}
-  },
-  "gotchas": [
-    "Flutterwave signs webhook payloads with HMAC-SHA256 using the secret hash configured in your dashboard. Compute HMAC-SHA256 of the raw request body and compare with the 'flutterwave-signature' header. Reject mismatches. Respond with HTTP 200 within 60 seconds — Flutterwave retries failed webhooks 3 times at 30-minute intervals.",
-    "The top-level event id field is 'webhook_id' for refund.completed (not 'id'), and 'timestamp' is a Unix epoch integer (not an ISO 8601 string). Flutterwave is inconsistent across event types — charge.completed historically uses 'id'.",
-    "data.status is broader than just 'completed' — it can be pending | requires_action | succeeded | failed | cancelled | completed | new. Always check the status explicitly before marking a refund as final.",
-    "Use the raw request body (bytes/string) when computing HMAC — re-serialising parsed JSON will reorder keys and break verification.",
-    "Use a constant-time comparison (e.g. node:crypto timingSafeEqual) when comparing signatures to avoid timing-side-channel attacks.",
-    "Deduplicate on webhook_id — Flutterwave may redeliver the same event after a transient failure on your side.",
-    "Always verify the refund server-side via GET /refunds/{id} before adjusting customer ledgers — never trust webhook data alone."
-  ]
-}
